Refactor loading and showing an ad to another class.
