+++ packages/app_open_example/lib/app_open_ad_manager.dart
@@ -1,0 +1,90 @@
+// Copyright 2021 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// ignore_for_file: public_member_api_docs
+
+import 'package:google_mobile_ads/google_mobile_ads.dart';
+
+/// Utility class that manages loading and showing app open ads.
+class AppOpenAdManager {
+
+  /// Maximum duration allowed between loading and showing the ad.
+  final Duration maxCacheDuration = Duration(hours: 4);
+
+  /// Keep track of load time so we don't show an expired ad.
+  DateTime? _appOpenLoadTime;
+
+  AppOpenAd? _appOpenAd;
+  bool _isShowingAppOpenAd = false;
+
+
+  /// Load an [AppOpenAd].
+  void loadAppOpenAd() {
+    AppOpenAd.load(
+      adUnitId: 'ca-app-pub-3940256099942544/3419835294',
+      orientation: AppOpenAd.orientationPortrait,
+      request: AdRequest(),
+      adLoadCallback: AppOpenAdLoadCallback(onAdLoaded: (ad) {
+        print('$ad loaded');
+        _appOpenLoadTime = DateTime.now();
+        _appOpenAd = ad;
+      }, onAdFailedToLoad: (error) {
+        print('AppOpenAd failed to load: $error');
+      },),
+    );
+  }
+
+  /// Shows the ad, if one exists and is not already being shown.
+  ///
+  /// If the previously cached ad has expired, this just loads and caches a
+  /// new ad.
+  void showAdIfAvailable() {
+    if (_appOpenAd == null || _appOpenLoadTime == null) {
+      print('Tried to show app open ad before it was loaded!');
+      return;
+    }
+    if (_isShowingAppOpenAd) {
+      print('Ad is already being shown.');
+      return;
+    }
+    if (DateTime.now().subtract(maxCacheDuration).isAfter(_appOpenLoadTime!)) {
+      print('Maximum cache duration exceeded. Loading another ad.');
+      _appOpenAd!.dispose();
+      _appOpenAd = null;
+      loadAppOpenAd();
+      return;
+    }
+    // Set the fullScreenContentCallback and show the ad.
+    _appOpenAd!.fullScreenContentCallback = FullScreenContentCallback(
+      onAdShowedFullScreenContent: (ad) {
+        _isShowingAppOpenAd = true;
+        print('$ad onAdShowedFullScreenContent');
+      },
+      onAdFailedToShowFullScreenContent: (ad, error) {
+        print('$ad onAdFailedToShowFullScreenContent: $error');
+        _isShowingAppOpenAd = false;
+        ad.dispose();
+        _appOpenAd = null;
+      },
+      onAdDismissedFullScreenContent: (ad) {
+        print('$ad onAdDismissedFullScreenContent');
+        _isShowingAppOpenAd = false;
+        ad.dispose();
+        _appOpenAd = null;
+        loadAppOpenAd();
+      },
+    );
+    _appOpenAd!.show();
+  }
+}
