+++ packages/app_open_example/lib/main.dart
@@ -14,6 +14,7 @@
 
 // ignore_for_file: public_member_api_docs
 
+import 'package:app_open_example/app_open_ad_manager.dart';
 import 'package:flutter/material.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 
@@ -50,22 +51,16 @@ class MyHomePage extends StatefulWidget {
 /// Overrides [WidgetsBindingObserver.didChangeAppLifecycleState]
 /// to handle showing an app open ad when the app is resumed.
 class _MyHomePageState extends State<MyHomePage> with WidgetsBindingObserver {
-  /// Maximum duration allowed between loading and showing the ad.
-  final Duration maxCacheDuration = Duration(hours: 4);
-
-  /// Keep track of load time so we don't show an expired ad.
-  DateTime? _appOpenLoadTime;
-
-  AppOpenAd? _appOpenAd;
-  bool _isShowingAppOpenAd = false;
 
   int _counter = 0;
+
+  AppOpenAdManager _appOpenAdManager = AppOpenAdManager();
 
   @override
   void initState() {
     super.initState();
     WidgetsBinding.instance!.addObserver(this);
-    _loadAppOpenAd();
+    _appOpenAdManager.loadAppOpenAd();
   }
 
   @override
@@ -73,63 +68,8 @@ class _MyHomePageState extends State<MyHomePage> with WidgetsBindingObserver {
     // Try to show an app open ad if the app is being resumed and
     // we're not already showing an app open ad.
     if (state == AppLifecycleState.resumed) {
-      _showAdIfAvailable();
+      _appOpenAdManager.showAdIfAvailable();
     }
-  }
-
-  void _loadAppOpenAd() {
-    AppOpenAd.load(
-      adUnitId: 'ca-app-pub-3940256099942544/3419835294',
-      orientation: AppOpenAd.orientationPortrait,
-      request: AdRequest(),
-      adLoadCallback: AppOpenAdLoadCallback(onAdLoaded: (ad) {
-        print('$ad loaded');
-        _appOpenLoadTime = DateTime.now();
-        _appOpenAd = ad;
-      }, onAdFailedToLoad: (error) {
-        print('AppOpenAd failed to load: $error');
-      }),
-    );
-  }
-
-  /// Show the ad, if one exists and is not already showing.
-  void _showAdIfAvailable() {
-    if (_appOpenAd == null || _appOpenLoadTime == null) {
-      print('Tried to show app open ad before it was loaded!');
-      return;
-    }
-    if (_isShowingAppOpenAd) {
-      print('Ad is already being shown.');
-      return;
-    }
-    if (DateTime.now().subtract(maxCacheDuration).isAfter(_appOpenLoadTime!)) {
-      print('Maximum cache duration exceeded. Loading another ad.');
-      _appOpenAd!.dispose();
-      _appOpenAd = null;
-      _loadAppOpenAd();
-      return;
-    }
-    // Set the fullScreenContentCallback and show the ad.
-    _appOpenAd!.fullScreenContentCallback = FullScreenContentCallback(
-      onAdShowedFullScreenContent: (ad) {
-        _isShowingAppOpenAd = true;
-        print('$ad onAdShowedFullScreenContent');
-      },
-      onAdFailedToShowFullScreenContent: (ad, error) {
-        print('$ad onAdFailedToShowFullScreenContent: $error');
-        _isShowingAppOpenAd = false;
-        ad.dispose();
-        _appOpenAd = null;
-      },
-      onAdDismissedFullScreenContent: (ad) {
-        print('$ad onAdDismissedFullScreenContent');
-        _isShowingAppOpenAd = false;
-        ad.dispose();
-        _appOpenAd = null;
-        _loadAppOpenAd();
-      },
-    );
-    _appOpenAd!.show();
   }
 
   void _incrementCounter() {
