+++ packages/google_mobile_ads/ios/Classes/FLTAd_Internal.m
@@ -368,10 +368,10 @@
                           return;
                         }
                         [weakSelf.manager
-                          onPaidEvent:weakSelf
-                                value:[[FLTAdValue alloc] initWithValue:value.value
-                                                              precision:(NSInteger)value.precision
-                                                           currencyCode:value.currencyCode]];
+                            onPaidEvent:weakSelf
+                                  value:[[FLTAdValue alloc] initWithValue:value.value
+                                                                precision:(NSInteger)value.precision
+                                                             currencyCode:value.currencyCode]];
                       };
 
                       [self.manager onAdLoaded:self responseInfo:ad.responseInfo];
@@ -456,10 +456,10 @@
                     return;
                   }
                   [weakSelf.manager
-                    onPaidEvent:weakSelf
-                          value:[[FLTAdValue alloc] initWithValue:value.value
-                                                        precision:(NSInteger)value.precision
-                                                     currencyCode:value.currencyCode]];
+                      onPaidEvent:weakSelf
+                            value:[[FLTAdValue alloc] initWithValue:value.value
+                                                          precision:(NSInteger)value.precision
+                                                       currencyCode:value.currencyCode]];
                 };
 
                 self->_insterstitial = ad;
@@ -544,10 +544,10 @@
                       return;
                     }
                     [weakSelf.manager
-                      onPaidEvent:weakSelf
-                            value:[[FLTAdValue alloc] initWithValue:value.value
-                                                          precision:(NSInteger)value.precision
-                                                       currencyCode:value.currencyCode]];
+                        onPaidEvent:weakSelf
+                              value:[[FLTAdValue alloc] initWithValue:value.value
+                                                            precision:(NSInteger)value.precision
+                                                         currencyCode:value.currencyCode]];
                   };
                   rewardedAd.fullScreenContentDelegate = self;
                   self->_rewardedView = rewardedAd;
